refactor(graduate): admin 사이드에서만 사용되는 일부 컴포넌트 이동 및 일부 페이지 로직 분리
