+++ apps/graduate/src/admin/pages/filePreview/utils/transformThesisFileResponse.ts
@@ -1,6 +1,6 @@
-import { ThesisFileResponse } from '~/shared/types';
-
 import { FileItem } from '../model/filePreview';
+
+import { ThesisFileResponse } from '~/admin/shared/types/file';
 
 export default function transformThesisResponse(
   data: ThesisFileResponse,
