+++ apps/graduate/src/admin/pages/schedule/hooks/useUpdateSchedule.ts
@@ -1,9 +1,10 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import type { Dayjs } from 'dayjs';
 
-import { updateSchedule } from '~/shared/api';
 import { KEYS, DATE_FORMAT } from '~/shared/constants';
 import { UpdateScheduleRequest } from '~/shared/types';
+
+import { updateSchedule } from '~/admin/shared/api';
 
 interface UseUpdateScheduleParams {
   scheduleId: number;
