+++ apps/graduate/src/admin/shared/api/notice.ts
