+++ apps/graduate/src/admin/shared/api/removeGraduationUser.ts
