+++ apps/graduate/src/admin/shared/api/removeGraduationUsersBatch.ts
