+++ apps/graduate/src/admin/shared/api/schedule.ts
