--- apps/graduate/src/admin/shared/api/student/index.ts
@@ -1,6 +1,0 @@
-export * from './fetchAdminUsers';
-export * from './fetchGraduationUsers';
-export * from './fetchGraduationUsersExcel';
-export * from './removeGraduationUser';
-export * from './removeGraduationUsersBatch';
-export * from './updateGraduationUsersBatchApprove';
