+++ apps/graduate/src/admin/shared/api/updateGraduationUsersBatchApprove.ts
