+++ apps/graduate/src/admin/shared/components/index.ts
@@ -1,0 +1,3 @@
+export * from './TextEditor';
+export * from './DataTable';
+export * from './Toolbar';
