+++ apps/graduate/src/admin/shared/hooks/index.ts
@@ -1,2 +1,6 @@
 export * from './admin-table';
 export * from './student';
+
+export { useScheduleList } from './useScheduleList';
+export { useScheduleDetail } from './useScheduleDetail';
+export { useScheduleContent } from './useScheduleContent';
