+++ apps/graduate/src/admin/shared/hooks/useScheduleContent.ts
