+++ apps/graduate/src/admin/shared/hooks/useScheduleDetail.ts
