+++ apps/graduate/src/admin/shared/types/file.ts
