+++ apps/graduate/src/shared/api/index.ts
@@ -1,4 +1,1 @@
 export * from './axios';
-export * from './schedule';
-export * from './notice';
-export * from './file';
