+++ apps/graduate/src/shared/components/index.ts
@@ -1,10 +1,7 @@
 export * from './Button';
-export * from './Toolbar';
 export * from './Header';
 export * from './Pagination';
-export * from './DataTable';
 export * from './Section';
-export * from './TextEditor';
 export { default as ProtectedRoute } from './ProtectedRoute';
 export { default as LoadingState } from './LoadingState';
 export { default as ErrorState } from './ErrorState';
