+++ apps/graduate/src/shared/types/index.ts
@@ -1,5 +1,4 @@
 export type * from './schedule';
 export type * from './notice';
 export type * from './graduationUser';
-export type * from './file';
 export * from './graduation';
